# Code/DotNet/APIDOCDemo/APIDOCDemo/Model/Event.cs
using System;

namespace APIDOCDemo.Model
{
    public class Event
    {
        public int Id { get; set; }
        public string Organiser { get; set; }
        public DateTime EventDate { get; set; }
    }
}