# Code/DotNet/APIDemoApplication/APIDemoApplication/Model/Event.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace APIDemoApplication.Model
{
    public class Event
    {
        public int Id { get; set; }
        public string EventName { get; set; }
        public DateTime EventDate { get; set; }
        public string Organiser { get; set; }
        public string Venue { get; set; }
    }
}