# Code/DotNet/APIDemoApplication/EventClient/event-swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"Events API"},"paths":{"/api/Events":{"get":{"tags":["Events"],"operationId":"ApiEventsGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/Event"}}}}},"post":{"tags":["Events"],"operationId":"ApiEventsPost","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"event","in":"body","required":false,"schema":{"$ref":"#/definitions/Event"}}],"responses":{"201":{"description":"Success","schema":{"$ref":"#/definitions/Event"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/ProblemDetails"}},"0":{"schema":{"$ref":"#/definitions/ProblemDetails"}}}}},"/api/Events/{id}":{"get":{"tags":["Events"],"operationId":"ApiEventsByIdGet","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ProblemDetails"}},"0":{"schema":{"$ref":"#/definitions/ProblemDetails"}}}},"put":{"tags":["Events"],"operationId":"ApiEventsByIdPut","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"event","in":"body","required":false,"schema":{"$ref":"#/definitions/Event"}}],"responses":{"204":{"description":"Success"},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ProblemDetails"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/ProblemDetails"}},"0":{"schema":{"$ref":"#/definitions/ProblemDetails"}}}},"delete":{"tags":["Events"],"operationId":"ApiEventsByIdDelete","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/Event"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/ProblemDetails"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/ProblemDetails"}},"0":{"schema":{"$ref":"#/definitions/ProblemDetails"}}}}}},"definitions":{"Event":{"type":"object","properties":{"id":{"format":"int32","type":"integer"},"eventName":{"type":"string"},"eventDate":{"format":"date-time","type":"string"},"organiser":{"type":"string"},"venue":{"type":"string"}}},"ProblemDetails":{"type":"object","properties":{"type":{"type":"string"},"title":{"type":"string"},"status":{"format":"int32","type":"integer"},"detail":{"type":"string"},"instance":{"type":"string"}},"additionalProperties":{"type":"object"}}}}