# Code/DotNet/APITest/APITest/Model/Event.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace APITest.Model
{
    public class Event
    {
        public int Id { get; set; }
        public string Organiser { get; set; }
        public string Venue { get; set; }
    }
}